Move client to domain
